requirements.txt
anthropic>=0.7.0
PyYAML>=6.0
python-dotenv>=1.0.0
